+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Add error aggregation. An exception raised during the dumper runtime will no
   longer halt execution. Instead, all errors are displayed after runtime.
+- Display real-time dumper status during dump execution.
 
 ## 4.1.0 (2022-10-18)